+++ api/lib/application/authentication/index.js
@@ -51,28 +51,6 @@ const register = async function (server) {
     },
     {
       method: 'POST',
-      path: '/api/token/anonymous',
-      config: {
-        auth: false,
-        payload: {
-          allow: 'application/x-www-form-urlencoded',
-        },
-        validate: {
-          payload: Joi.object().required().keys({
-            campaign_code: Joi.string().required(),
-            lang: Joi.string().required(),
-          }),
-        },
-        handler: AuthenticationController.authenticateAnonymousUser,
-        notes: [
-          "- Cette route permet de créer un utilisateur à partir d'un code parcours Accès Simplifié\n" +
-            "- Elle retournera un access token Pix correspondant à l'utilisateur.",
-        ],
-        tags: ['api'],
-      },
-    },
-    {
-      method: 'POST',
       path: '/api/revoke',
       config: {
         auth: false,
