+++ api/tests/acceptance/application/authentication/authentication-route_test.js
@@ -103,96 +103,6 @@ describe('Acceptance | Controller | authentication-controller', function () {
     });
   });
 
-  describe('POST /api/token/anonymous', function () {
-    let server;
-    let options;
-
-    beforeEach(async function () {
-      server = await createServer();
-    });
-
-    context('When is not simplified Access Campaign', function () {
-      const campaignCode = 'RANDOM123';
-      const lang = 'en';
-
-      beforeEach(async function () {
-        const targetProfile = databaseBuilder.factory.buildTargetProfile({ isSimplifiedAccess: false });
-        databaseBuilder.factory.buildCampaign({ code: campaignCode, targetProfile });
-
-        options = {
-          method: 'POST',
-          url: '/api/token/anonymous',
-          headers: {
-            'content-type': 'application/x-www-form-urlencoded',
-          },
-          payload: querystring.stringify({
-            campaign_code: campaignCode,
-            lang,
-          }),
-        };
-
-        await databaseBuilder.commit();
-      });
-
-      it('should return an 401', async function () {
-        // when
-        const response = await server.inject(options);
-
-        // then
-        expect(response.statusCode).to.equal(401);
-        expect(response.result.errors[0].detail).to.equal("L'utilisateur ne peut pas être créé");
-      });
-    });
-
-    context('When is simplified Access Campaign', function () {
-      const simplifiedAccessCampaignCode = 'SIMPLIFIE';
-      const firstName = '';
-      const lastName = '';
-      const isAnonymous = true;
-      const lang = 'en';
-
-      beforeEach(async function () {
-        const targetProfileId = databaseBuilder.factory.buildTargetProfile({ isSimplifiedAccess: true }).id;
-        databaseBuilder.factory.buildCampaign({ code: simplifiedAccessCampaignCode, targetProfileId });
-
-        options = {
-          method: 'POST',
-          url: '/api/token/anonymous',
-          headers: {
-            'content-type': 'application/x-www-form-urlencoded',
-          },
-          payload: querystring.stringify({
-            campaign_code: simplifiedAccessCampaignCode,
-            lang,
-          }),
-        };
-
-        await databaseBuilder.commit();
-      });
-
-      it('should return a 200 with accessToken', async function () {
-        // when
-        const response = await server.inject(options);
-        const result = response.result;
-
-        // then
-        expect(response.statusCode).to.equal(200);
-
-        expect(result.token_type).to.equal('bearer');
-        expect(result.access_token).to.exist;
-      });
-
-      it('should create an anonymous user', async function () {
-        // when
-        await server.inject(options);
-
-        // then
-        const users = await knex('users').where({ firstName, lastName, isAnonymous });
-        expect(users[0]).to.exist;
-      });
-    });
-  });
-
   describe('POST /api/application/token', function () {
     let server;
     let options;
